Extract gamepad cheatsheet view
